requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
filelock==3.4.0
idna==3.3
markdown2==2.4.2
Pillow==9.0.1
platformdirs==2.4.0
psycopg2==2.9.3
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
virtualenv==20.10.0